moved cache boolean to app.js, default false

+++ api/app.js
@@ -17,8 +17,14 @@ db = require('nano')(DBconfig.protocol+
   '@'   + DBconfig.host + 
   ':'   + DBconfig.port + 
   '/'   + DBconfig.database);
+
+DEBUG = false;
+CACHE = false; //not fully implemented
+
+if (process.argv.indexOf('debug')    !== -1) DEBUG = true;
+if (process.argv.indexOf('no-cache') !== -1) CACHE = false; 
   
-redis = require("redis").createClient();
+if (CACHE) redis = require("redis").createClient();
   
 gatewayList = require('./gateways.json');
   // TODO find permanent location for gateways list
@@ -79,10 +85,14 @@ function requestHandler(req, res) {
   } 
 }
 
-redis.flushdb(); //reset cache on restart
-redis.on("error", function (err) {
-  console.log("Error " + err);
-});
+if (CACHE) {
+  redis.flushdb(); //reset cache on restart
+  redis.on("error", function (err) {
+    console.log("Error " + err);
+    CACHE = false; //turn it off if its not workings
+  });  
+}
+
